[cosmos] Follow protocol definition of "fees"
The "fee" object in transactions actually provides the max amount of
fees the user is ready to pay for a transaction. There is no meaning to
multiplying this value with the given "gas" value (which is the gas
limit).

The only meaningful operation with those 2 values is to compute the
gasPrice using fee / gas, but this should only be provided as a
Transaction API const method, _fee and _gas should keep their meaning of
"transaction fees" and "gas limit" respectively