[cosmos] Type overhaul. use ledger::core::cosmos as serialized database
cosmos::Block should be the only Block with data in all code,
same for cosmos::Transaction, cosmos::Account etc.

The goal is to use composition to provide the necessary APIs in the
djinni interface concret types.

Also, stop using ledger::core::Block which is deprecated, and start
using ledger::core::api::Block instead (again, this is only a change to
cosmos::Block typedef, DO NOT use anything else in the code)